requirements.txt
langchain-groq
langchain_community
faiss-cpu
flask
